WIP Refactor - Create a concensus-sdk to normalize a facade to perform operations on Poll / Vote entities via either a REST API bindings or Web3 bindings
