+++ lib/api/firebase/index.js
