+++ lib/api/registerForPushNotificationsAsync.js
