+++ lib/concensus-sdk/index.js
@@ -1,0 +1,9 @@
+import Poll from './rest-api/poll';
+import Vote from './rest-api/vote';
+
+const concensusSdk = {
+  Poll,
+  Vote
+};
+
+export default concensusSdk;
