+++ lib/concensus-sdk/rest-api/vote.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+const Vote = {
+  delete: function () {
+    axios.delete('http://4d23f078.ngrok.io/votes');
+  }
+};
+
+export default Vote;
