+++ screens/PollScreen.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import axios from 'axios';
 import { Dimensions, StyleSheet, ScrollView, Text, View } from 'react-native';
 import PollComponent from '../components/PollComponent.js';
 import VoteComponent, { VOTE_ENUM } from '../components/VoteComponent.js';
 import DiscussionSection from '../components/DiscussionSection.js';
-import { getUserVote } from '../api/firebase';
+import { getUserVote } from '../lib/api/firebase';
+import { Vote } from '../lib/concensus-sdk';
 
 function msToTime(s) {
     let secs = s % 60;
@@ -59,7 +59,7 @@ class PollScreen extends React.Component {
             this.setState(previousState => {
                 console.log('previousState', previousState.timeRemaining);
                 if (previousState.timeRemaining < 1) {
-                    axios.delete('http://4d23f078.ngrok.io/votes');
+                    Vote.delete();
                 }
                 return { timeRemaining: previousState.timeRemaining - 1 };
             });
